rename constants
